Implement support for non-interactive sessions
In non-interactive sessions (i.e. running podman without "-i") stdin
is not pollable, which triggers an error when trying to register the
event in event_manager.

Add a method in event_manager to be able to check if a file descriptor
is pollable, and use it to determine if stdin can be polled. If that's
not the case, turn off the console's interactive mode, to avoid
reporting stdin as file descriptor of interest.

Signed-off-by: Sergio Lopez <slp@redhat.com>