vmm: Split the kernel command line in prolog and epilog
The kernel command line is user configurable, but the VMM also touches
for adding references to configured virtio devices. To ensure the user
can put data before and after the automatically added content, split
up the kernel command line properties to add one for the epilog and
another for the prolog.

Signed-off-by: Sergio Lopez <slp@redhat.com>