devices/fs: Implement support for DAX
Implement support for DAX in virtio-fs. We use a static 512MB window
at the end of the RAM. This enables the guest to bypass its own cache
for some VFS operations, reducing the amount of memory used in the
guest for its own FS cache, and increasing the number of pages we can
return to the host to free them.

This feature is supported in Linux/KVM, and has been tested on x86_64
and Aarch64. We can't use DAX on macOS, because Hypervisor.framework
doesn't allow adding MAP_SHARED maps to the VM (or, at least, I
couldn't find a way).

Signed-off-by: Sergio Lopez <slp@redhat.com>