SEV: Load qboot into the reset vector
In the regular version of libkrun, the integrated VMM starts booting
directly in 64 bits and into the kernel entry point, after setting up
some initial data structures (page tables, e820, mptables...).

With SEV, those data structures are problem, as they become part of
the launch measurement.

So, if targeting SEV, use qboot as firmware, loading it into the
reset, and don't set those initial data structures (they'll be set by
qboot after the VM is started).

This commit also introduces the "amd-sev" feature to allow users to
choose this flavor or the library.

Signed-off-by: Sergio Lopez <slp@redhat.com>