SEV: Add support for initramfs
With virtio-fs gone, we need a trusted place to put the "init" binary
that sets up the environment in the guest for running the workload
entry point and, in case of using LUKS, to unlock the root file
system.

The obvious replacement is using an initramfs. This change extends
libkrun to load the initrd bundled in the SEV variant of libkrunfw.

Signed-off-by: Sergio Lopez <slp@redhat.com>