init: override kernel's HOME and TERM envs
Starting 3ac06634ca56b35b91f7329fd686d97678574b46, init.c is able to
read the workload parameters from a configuration file. After this
change, parameters passed as environment variables still take
preference over the ones specified in the config file, to allow our
callers to easily override them without altering said file.

The problem with this approach is that the kernel unconditionally
exports the HOME and TERM variables, so the ones in the config file
are never applied.

In this change, we always override HOME and TERM with the values in
the config file, and we introduce KRUN_HOME and KRUN_TERM to allow our
callers to still be able to override them using environment variables.

Signed-off-by: Sergio Lopez <slp@redhat.com>