tee: Make TeeConfig required
When running a workload in a TEE environment, a TeeConfig should be
required. It doesn't make much sense to run a workload in a
TEE-environment yet not give a configuration of said TEE. This is
especially true now, as libkrun will be able to ship support for both
SEV and SEV-SNP in one binary.

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>
Signed-off-by: Sergio Lopez <slp@redhat.com>