examples/snp: Include use an unencrypted disk image to get TEE config
For SNP, we don't care much for krun_set_tee_config(), as the
attestation is done from the guest. Instead, wrtie the TEE config file
to the data disk and allow the guest to read it from the init binary.

Also including an example TEE config disk image
"snp-config-data-disk.img" with the workload ID "test" and attestation
URL of localhost to allow for a user to test the attestation
functionality.

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>