init/snp: Modify attestation report structure + JSON encoding
With firmware update 1.52, the firmware required extra padding bits to
align properly. Furthermore, instead of traversing each field and
writing it out to the JSON buffer, use OpenSSL to hex encode the report
before sending to the attestation server.

Due to the fact that we now submit the SNP generation (which allows the
attestation server to retrieve the certificate chain), we don't need to
send the certificate chain to the attestation server from libkrun. Thus,
only grab the report using SNP_GET_REPORT instead of SNP_GET_EXT_REPORT.

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>