passt: properly handle multiple consecutive partial writes to socket
Previously when write_loop was called by try_finish_write and it wrote
something to the socket (but not fully finish the current frame),
it would set last_partial_write_length to an incorect value.

This commit also changes the type of last_partial_write_length to usize
from, Option<NonZeroUsize> because it seems to be simpler to work with.

Signed-off-by: Matej Hrica <mhrica@redhat.com>