[Linux] Make krun_start_enter error when root directory does not exist
Previously krun_start_enter would succeed and the guest kernel would
just panic. The root filesystem directory was opened lazily when
the guest kernel used fuse init opcode.
This commit changes it so the root directory is opened when
creating the fs device.

This only fixes it on Linux, but not in the macOS implementation.

Signed-off-by: Matej Hrica <mhrica@redhat.com>