lib: Add vsock port map and API for inserting into map
A VM guest's vsock UNIX port map will allow the guest to map a vsock
port to a UNIX socket in the host. In this situation, a host will
start a service listening on a UNIX socket with a known path. The host
can then use krun_add_vsock_port() to add a vsock->UNIX entry into the
guest, allowing the guest to connect to then UNIX socket via the mapped
port.

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>