devices/gpio: import a PL061-based GPIO
Import a PL061-based GPIO to have an interface in the guest to send our
shut down signal. This device, along with its FDT entry, will instruct
Linux to enable a gpio-keys interface to receive key events.

When a consumer of the library writes to the eventfd received from
krun_get_shutdown_eventfd(), this device will send a key press event
followed by a key release event, for the key code KEY_RESTART. We're
using KEY_RESTART instead of KEY_POWEROFF because it's easier to
identify a vCPU exiting for a reboot than for a poweroff.

For the moment we're only implementing support on macOS/aarch64, but
should be possible to work mostly in the same way on Linux/aarch64.

Signed-off-by: Sergio Lopez <slp@redhat.com>