virtio/fs/macos: overhaul to use macos inodes
The hardest problem of virtio-fs is keeping a relationship between
inodes in virtio-fs (guest's perspective) and inodes in the backing
filesystem (host's perspective).

On Linux, this forces us to keep a fd open per each alive dirent in the
guest (hence the "forget/batch_forget" logic). Trying to use the same
strategy of macOS proved to be very difficult as the process can easily
reach a limit in the number of open files that can't be easily
increased.

To overcome this difficuly, we were using a complex combination of both
fd and paths caches. In this commit, we overhaul the whole
implementation to remove the use of those caches to take advange instead
of macOS ability to access inodes directly through "/.vol".

This strategy simplifies the code, reduces the cost of each operation
and makes the implementation significantly more robust. Its main
drawback is that we can't support traversing mutiple filesystems through
the same mountpoint on the host, but that's a reasonable limitation on
macOS that shouldn't affect the vast majority of users.

Signed-off-by: Sergio Lopez <slp@redhat.com>