virtio/block: support large segment numbers
virtio-block can support large segment numbers (visible in Linux as
"max_segments") by putting more than one data descriptor in the
descriptor chain. So far, we only supported a single data descriptor,
which on Linux implies "max_segments == 1".

In this commit we reuse the Reader/Writer abstractions that have been
used by virtio/fs for very long to be able to easily digest large
descriptor chains, and we extend VirtioBlkConfig to declare that we
support up to 254 (QUEUE_SIZE - 2) segments.

This provides a huge performance boost for both reading and writing
large chunks of data.

While there, also implement support for EVENT_IDX to reduce the traffic
of signals between the guest and the host.

Signed-off-by: Sergio Lopez <slp@redhat.com>