Add SMBIOS internal crate
This crate will build and write in memory all the required
tables for SMBIOS. It build tables type 0, 1 and 11 if
provided.

It's not SMBIOS compliant, since it only set 2 of the 10
required tables. It also doesn't set the UUID in type 1 table
(i.e., system information), that is not clear it's required
in the spec (QEMU set a 0 UUID if not provided).

The type 0 characteristics bits follow the example in QEMU
and EDK2

Signed-off-by: German Maglione <gmaglione@redhat.com>