virtio/fs: Fix corrupted dirents due to trailing garbage
The dirent name length is supposed to be the actual strlen() of the
name, without any padding. SYS_getdents64, however, will pad the records
it returns (for alignment, etc.) as well as NUL-terminate the name.

This was causing dirents in the guest kernel to have spurious NUL
padding treated as part of the name, which in turn was breaking getcwd()
under some conditions (and probably other more subtle things).

Signed-off-by: Asahi Lina <lina@asahilina.net>