virtio/fs/macos: decouple host-guest inodes
Exposing the host inode to the guest was a nice optimization that
also simplified the paths to lookup and open and inode, but it also
means we can't properly support submount (which, albeit unusual on
macOS, they can be there).

Go back to decoupling host-guest inode, like the implementation for
Linux, but keep using "/.vol" to open the handles to avoid having to
keep them open while the entry in guest's dirent is alive.

Signed-off-by: Sergio Lopez <slp@redhat.com>