arch: implement SHM region management
We have multiple devices (fs and gpu) that can use SHM regions. So far,
we were creating a single SHM region so only one device could make use
of it. In this commit, we implement SHM region management so multiple
devices can request their own regions.

A perhaps controversial decision is to NOT use the GuestMemory
abstraction to manage the SHM regions, and instead operate them naked.
The reason behind this is that GuestMemory is primarily designed with
the intention of self-managing the memory mappings, which is exactly the
opposite of what we need to do with SHM regions.

Also, SHM regions operate in very different ways on Linux/KVM and
macOS/HVF. On the first, we need to create a large mapping, register it
with KVM, and then lay SHM windows on top of it as nested mappings. On
the latter, nested mappings aren't supported, so each SHM window must be
explicitly registered with HVF, so we don't need to create nor register
an initial large mapping.

In addition to this, SHM windows on Linux/KVM operate with offsets on
top of the host address, while macOS/HVF operate with offsets on top of
the guest address.

For the moment, we're hardcoding the SHM region sizes for gpu and fs. A
future commit will extend the API so users can configure those sizes as
desired.

Signed-off-by: Sergio Lopez <slp@redhat.com>