virtio/fs/linux: Fix xattrs on symlinks
The f*attr library of functions only works on files opened normally.
Symlinks can only be opened with O_PATH, but that doesn't support xattr
operations. So, to handle symlinks, we need to use the l*attr family of
functions directly on the pathname.

Thankfully, this does still work via /proc/self/fd/* for fds opened with
O_PATH, so we can still do fd-relative operations (l*attr on such a path
resolves the magic symlink and operates on the symlink that is actually
opened as that fd).

Signed-off-by: Asahi Lina <lina@asahilina.net>