virtio/fs: Add support for exporting FDs to other subsystems
In order to support fence passing for the X11 protocol, we need to be
able to share an opened fd (to a /dev/shm file opened with DAX) between
the guest and the host. To do this, introduce a new ioctl() operation on
the filesystem, which tags the handle as exported and inserts it into a
shared table that can be shared with other subsystems.

Semantics:
- Shared FDs are required to be kept open by the guest until the host
  code uses the share. Closing the fd will remove the share (this
  ensures fds/shared don't leak if a share is unused).
- Shares are identified by a filesystem ID and a handle ID, both u64.
  They are opaque as far as the user is concerned. Filesystem IDs are
  arbitrarily assigned at creation time from a static incrementing
  atomic counter.
- There is no particular permissions/privilege checking for this
  mechanism. In other words, once an FD is exported, it can be used in
  any context where exported FDs are expected by handle (which is
  predictable and not secure), with no regard to permissions or
  privilege. It is expected that this mechanism will be used with a
  single-privilege-domain VM scenario.

Only Linux is supported at this time (macOS only has stub support for
the fields, no implemetation).

Signed-off-by: Asahi Lina <lina@asahilina.net>