testing: use qboot-krunfw instead of TDVF
Feeding libkrun TDVF as the firmware results in a failed assertion in
edk2 because libkrun doesnt' support PCI or ACPI, which are both
required for edk2 to work properly. Therefore, we should try to use
qboot instead.

Currently, qboot is starting the VM boot process, but it will triple
fault once control gets given to the vCPU. I theorize this is because
there isn't anything in the hypervisor or the firmware that knows how to
deal with a #VE (Virtualization Exception). Therefore, the exception
can't get handled and causes a double, then triple, fault.

Signed-off-by: Jake Correnti <jakecorrenti+github@proton.me>