Intruduce an E2E testing framework
Add a tests directory containing an E2E testing framework.
The tests are E2E in the sense that they use the public API of libkrun to
start a VM and run a test program in the guest.
Note that currently this is very limited, because there are no other userspace
executables or libraries (not even libc) in the guest apart from `guest-agent`.

`guest-agent` is a statically linked Rust executable, that executes the guest
part of each test.

In the future this can be extended to run tests with specific images which would
allow the use of system libraries by the test program.

This also introduces 2 tests:
test_vm_config - asserts the VM is constructed with the correct number of CPUs
                 and amount of memory. This also serves as an example how the
		 tests can be parameterized.
test_vsock_guest_connect - This tests the guest connecting to a vsock port
                           created by the `krun_add_vsock_port` API.

Signed-off-by: Matej Hrica <mhrica@redhat.com>