Makefile: Make libkrun.pc target .PHONY and generate it on install
The install target should depend on generation of the libkrun.pc file,
because it installs it.
The libkrun.pc target should be .PHONY because libkrun.pc file might
already exists and be generated, but it could be made using a different PREFIX,
which would lead to the installed libkrun.pc pointing to the wrong prefix.

Signed-off-by: Matej Hrica <mhrica@redhat.com>