macos/vstate: unify MPIDR generation
With HVF, we generate MPIDR by taking Vcpu.id and setting Aff1 to
that value by doing a left-shift, as this is what the GIC expects
to be found in the distributor.

But, in vstate::Vcpu::configure_aarch64, we were using the Vcpu.id
without left-shifting it.

For consistency, let's generate MPIDR for the Vcpu once and use it
everywhere.

Signed-off-by: Sergio Lopez <slp@redhat.com>