Share vm and vcpu fds
These descriptors are required for handling the KVM_FAULT_MEMORY vmexit.
Although this vmexit is only present when the tee feature is enabled,
changing for a reference does not seem the right thing to do by using
feratures. Use reference and protect it to be able to share these
descriptors. These changes prepare the code for handling
KVM_EXIT_FAULT_MEMORY

Signed-off-by: Matias Ezequiel Vara Larsen <mvaralar@redhat.com>