libkrun: avoid failing when libkrunfw is missing
In the last iteration of PR#256 we introduced the use of LazyLock,
which helps us load the library earlier. The implementation was
wronging because we were unwrapping if the library is missing,
which is exactly what we were trying to prevent.

Replace the value with an Option, which allows us to avoid the
unwrap and actually achieves the purpose of making libkrunfw
optional.

Signed-off-by: Sergio Lopez <slp@redhat.com>