vsock: delay listening for events in TCP socket
As indicated in 3bd3b98916eed4816ec19bc627c94edf87695090, when a client
in the guest attempts to send data to the server immediately after
connecting to it, we might receive data in the TCP socket before we're
able to actually process it.

To avoid this situation, delay setting EventSet::IN until the vsock
transport has been established in confirm_connect.

Signed-off-by: Sergio Lopez <slp@redhat.com>