devices/legacy: implement `BusDevice` trait for `IoApic`
Implement the `BusDevice` trait for `IoApic` to handle reads and writes
from the system bus.

When reading the `IO_WIN` register, it's important to verify the data is
only 32-bits in size, as the register is 64-bits, but needs to do two
consecutive reads.

Like `read`, `write` needs to verify the data is 32-bits so the
register can be read twice and get the full 64-bits of data. In `write`,
the IO APIC can't write to the Version or Arbitration regiseters,
because those are Read-Only registers on the device.

Signed-off-by: Jake Correnti <jakecorrenti+github@proton.me>