devices, virtio/gpu: Introduce a display subsystem
Introduce a DisplayBackend trait and a DisplayBackendNoop implementation. Hook
up the gpu device to use the display backend. For now this only hooks up the
simplest and slowest option of display output - the output is copied to the
CPU using the transfer_read operation and passed to the display backend. The
DisplayBackendNoop doesn't provide any displays, so this doesn't display
anything yet.

Introduce a new public API function krun_set_display() for configuring VM
displays.

The virtio-gpu code is based on rust-vmm/vhost-device-gpu.

Signed-off-by: Matej Hrica <mhrica@redhat.com>