Declare env_logger as dependency only in the top level crate
The correct way to use the env_logger crate is to only depend on it the toplevel
aplication crate. In other crates we should just use the `log` crate facade
(which we already do).

Drop the env_logger dependency from all of our internal crates, and just keep it
in the the `libkrun` crate.

Signed-off-by: Matej Hrica <mhrica@redhat.com>