devices: Refactor interrupt sending into MmioTransport
Create an abstraction InterruptTransport which sits at the MmioTransport layer.
This holds all the necessary fields to send an interrupt. The InterruptTransport
is passed to a VirtioDevice as a parameter for the activate() method call and is
saved inside DeviceState::Active.

Make IrqChip now passed into the MmioTransport::new. Because it is now always
required, this removes the need to check if it is set before sending an interrupt.

Previously we had around 15 copied implementations of signal_used_queue in the
project. For every implementation/place (worker threads, etc) where
signal_used_queue was necessary, one had to also pass around 4 variables.

Signed-off-by: Matej Hrica <mhrica@redhat.com>