hvf: mask out SME in ID_AA64PFR1_EL1
If SME is present in ID_AA64PFR1_EL1 in the VM, the guest will break
after enabling the MMU. I don't see an architectural reason for this to
happen, so I suspect it's a deliberate action of some macOS component to
avoid exposing the feature to VMs.

In any case, let's just mask out SME from ID_AA64PFR1_EL1 to fix nested
virt on M4 devices.

Signed-off-by: Sergio Lopez <slp@redhat.com>