Introduce APIs for hooking up display implementations
Introduce krun_set_display for specifying the amount of displays and their resoultion.
Introduce the the krun_set_display_backend for specifying the display backend as
a vtable and a bitmask of features. This also introduces a new header
libkrun_display.h and a krun_display rust crate which serves as a bidirectional
bindings for the C display API.

Signed-off-by: Matej Hrica <mhrica@redhat.com>