include: vmm: add `krun_set_serial_console` API
Currently, libkrun forces the user to manually change the source
code if they are not building with EFI=1 and want to use the
serial console and instead of the virtio-console.

Provide an API, `krun_set_serial_console` allowing the user to specify
whether the serial console should be enabled. The serial console remains
the default on EFI=1 use cases.

Signed-off-by: Jake Correnti <jakecorrenti+github@proton.me>