Vendor the krun-sys bindingings into this repository
Vendor the source code of krun-sys bindings from https://github.com/AsahiLinux/muvm
to this repository. We depend on krun-sys internally for our integration tests
and soon we may also have examples written in Rust. The idea is that this
repository will become the source of the crates.io crate.

Original authors are Teoh Han Hui <teohhanhui@gmail.com> and Sergio Lopez
<slp@redhat.com>.

Co-authored-by: Teoh Han Hui <teohhanhui@gmail.com>
Co-authored-by: Sergio Lopez <slp@redhat.com>
Signed-off-by: Matej Hrica <mhrica@redhat.com>