krun-sys: Expose feature for configuring libclang loading
`krun-sys` depends on the bindgen crate which in turn depends on clang-sys, but
`clang-sys` has a curious problem:
Let's say you are compiling crate `A`, which depends on crates `B` and `C`, which
both depend on `clang-sys` but with different features. If the crate `B` doesn't
use the "runtime" feature (`default-features=false`), but crate `C` does, then crate
`A` will fail (in build.rs) with the message:
`a libclang shared library is not loaded on this thread`.

To fix this all crates in a project that use clang-sys have to either enable the "runtime"
feature or disable it.

To work around this issue expose the bindgen/clang-sys feature as
"bindgen_clang_runtime" feature of krun-sys.

Some discussion about this issue:
https://github.com/twistedfall/opencv-rust/issues/462#issuecomment-1551927461

Signed-off-by: Matej Hrica <mhrica@redhat.com>