rutabaga_gfx: do not try to RW map write-sealed memfds from host Wayland compositor
mmap() returns EPERM if the fd has been sealed for (against?) the given mode.
One such case that would instantly cause sommelier to fail to work:
at least mutter sends the keymap as write-sealed, and the shm code was
hardcoded to RW access. Let's read the seals and unset the write bit when
necessary.

In upstream rutabaga_gfx, that case is avoided by the special-casing of
the Wayland keymap, with hardcoded read-only mode (but it's special-cased
for another reason, having to do with single- vs multi-use).

Signed-off-by: Val Packett <val@invisiblethingslab.com>