devices/vsock: always reply to TSI control messages
For TSI control messages that require an answer (connect, listen,
accept and getname), reply with an error message even if the proxy
is not present.

Until now, under normal circumstances, we should never receive one of
those requests without a proxy, but in the next commit we're going to
start rejecting the creation of proxies under certain circumstances, so
we need to do this first.

Signed-off-by: Sergio Lopez <slp@redhat.com>