devices/vsock: selectively enable AF_UNIX in TSI
AF_UNIX transparency is only supported when running in a containerized
environment. We detect such case by checking we only have a single
virtio-fs device, and that device is pointing to "/", which is the
scenario when running under podman+crun.

In libkrun 2.x, the whole network configuration, including the TSI
features, should be exposed to the users explicitly.

Signed-off-by: Sergio Lopez <slp@redhat.com>