Fix behavior of krun_add_virtio_console_default
When using krun_add_virtio_console_default with file descriptors that weren't
normal STDIN/STDOUT/STDERR we could get weird behavior that wasn't in line with
the behavior specified in the header.

This also makes sure to set the change the terminal mode actually associated with
the provided file descriptors instead of changing the terminal mode associated with
normal stdin/stdout/stderr.

Signed-off-by: Matej Hrica <mhrica@redhat.com>

fixups

Signed-off-by: Matej Hrica <mhrica@redhat.com>