init: Define specific initramfs for nitro enclaves
AWS nitro enclaves will follow a slightly different initramfs than
standard libkrun. Rather than following the confidential workloads
model, we'll result to receiving the archived rootfs for the system
from the hypervisor.

Perhaps this may at some point be applicable to other TEE architectures,
but for now it is only relevant to the nitro variant.

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>