nitro: Add configurable rootfs for enclaves
To avoid regenerating an EIF file within changing applications, re-use
the krun_set_root API to set the rootfs for a nitro enclave.

Although re-using this API, the behavior is slightly different in the
nitro variant. Rather than attaching a virtio-fs device (which is
currently unsupported in nitro enclaves), archive the rootfs and write
it over a vsock connected to the enclave. The enclave will read,
extract, and mount the rootfs that is read over the vsock.

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>