nitro: Use cached EIF file if none configured
With the underlying application decoupled from the EIF file, create and
cache a stock EIF file that can be used to run enclaves with.
Applications that want to run "standard" containzerized applications can
now do so without specifying an EIF file.

Update the nitro example to not require an EIF file.

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>