tests: Crate all test dirs under a single temp dir
Instead of creating a tmp dir with random suffix for each test, create a tmp dir
with the suffix for the whole run of the test. This allows the user to more
easily view results of a run under a single directory.

This also adds a --base-dir flag for overriding the test directory, and
--keep-all which doesn't delete any of the files generated by the test.

Signed-off-by: Matej Hrica <mhrica@redhat.com>