builder: check that kernel fits in memory
For the KernelCopy payload, check that the kernel really fits in memory
before copying it. We were counting on guest_mem.write() to fail if the
addresses are outside the covered regions, but this may not happen if
we've added shared memory regions to the guest memory. Checking against
arch_mem_info.ram_last_addr is more reliable.

Signed-off-by: Sergio Lopez <slp@redhat.com>