vmm: linux: Do not use guest_memfd outside of TEE builds (fixes #478)
aarch64 KVM has recently introduced guest_memfd support, but since there
are no VM TEEs with upstream support on that arch, it returns EINVAL when
we try to create a non-mappable memfd (which is what flags==0 implies).

Then.. even if we make it mappable and skip setting the private flag,
we get a 100% CPU busy loop in the guest. Same happens in QEMU though
so that's not a libkrun bug. Still, let's avoid *all* of that by not
trying to use guest_memfd for protecting the VM's memory outside of
platforms where that's explicitly supported (SEV/TDX).

Signed-off-by: Val Packett <val@invisiblethingslab.com>