nitro: improve console support
Rather than tightly coupling the application output with the kernel
output, which is only enabled when the debug start flag is passed to the
nitro enclave, separate the two.

The application output socket in the guest will always be created and
the host is expected to create a socket to connect to the guest and
listen for output.

The debug output socket can be optionally created by the host if it is
desired, no longer forcing the user to see the kernel messages.

Signed-off-by: Jake Correnti <jakecorrenti+github@proton.me>