nitro: Add writer for enclave arguments
To modularize the task of writing the enclave's arguments (rootfs, exec
path, etc) to the guets via vsock, create a writer specifically tasked
with writing the arguments to the enclave.

This will allow for better extensibility when other arguments (network,
storage, attestation credentials, etc) would like to be given to the
enclave's initramfs.

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>