Fix external kernel memory region calculation
The external kernel payload type was not passing the firmware size to
the architecture-specific memory region calculation function. This could
lead to incorrect memory layout when using external kernels with
firmware. The change ensures the firmware size is properly accounted for
in the memory region setup. This inconsistency caused VM creation to
fail with HV_DENIED (-85377017) on macOS when using direct kernel boot
with the Hypervisor.framework.

Signed-off-by: Eric Curtin <eric.curtin@docker.com>